requirements.txt
PyQt5>=5.15.0
qrcode>=7.3
Pillow>=9.0.0
